+++ packages/playground/src/components/MyAudio.tsx
@@ -1,0 +1,14 @@
+import { Audio, Div } from 'aware-components';
+
+export function MyAudio() {
+  return (
+    <Div>
+      <Audio autoPlay={true} controls>
+        <source src='horse.ogg' type='audio/ogg' />
+        <source src='horse.mp3' type='audio/mpeg' />
+        Your browser does not support the audio element
+      </Audio>
+      <p>Transcript: A horse sound</p>
+    </Div>
+  );
+}